child_process: add example for execFileSync method and ref to stdio
Added an example to the execFileSync method. This demonstrates how to
handle exceptions and access the stderr and stdio properties that are
attached to the Error object in a catch block.

Added a link to the detailed stdio section nested under
child_process.spawn() from each child_process sync method option
description.

Fixes: https://github.com/nodejs/node/issues/39306