test_runner: use validateStringArray for `timers.enable()`
`apis` which is argument of `timers.enable()` is string array.
So use `validatStringArray` instead of `validateArray`. And
`options` is optional, so update JSDoc.
In document, some default value of `timers` are missed such as
`setImmediate` and `clearImmediate`. Next, `milliseconds` which
is argument of `timers.tick() is optional and default is 1.